Menambahkan endpoint TemplateAktivitas, mekanisme cek luas lahan sebelum menambahkan aktivitas, dan memperbarui utility-utility yang dibutuhkan

+++ src/main/java/utils/DataPengguna.java
@@ -13,24 +13,27 @@ public class DataPengguna {
         Path lahanFilePath = folderPath.resolve("lahan.json");
         Path aktivitasFilePath = folderPath.resolve("aktivitas.json");
         Path faseFilePath = folderPath.resolve("fase.json");
+        Path templateAktivitasFilePath = folderPath.resolve("template_aktivitas.json");
         Path autoIncFilePath = folderPath.resolve("auto_inc");
 
         if (!(Files.exists(lahanFilePath) && Files.exists(aktivitasFilePath) && Files.exists(faseFilePath)
-                && Files.exists(autoIncFilePath))) {
+                && Files.exists(autoIncFilePath) && Files.exists(templateAktivitasFilePath))) {
             try {
                 if (!Files.exists(folderPath)) {
                     Files.createDirectories(folderPath);
                 }
 
+                Files.deleteIfExists(templateAktivitasFilePath);
                 Files.deleteIfExists(lahanFilePath);
                 Files.deleteIfExists(aktivitasFilePath);
                 Files.deleteIfExists(faseFilePath);
                 Files.deleteIfExists(autoIncFilePath);
 
+                Files.createFile(templateAktivitasFilePath);
                 Files.createFile(lahanFilePath);
                 Files.createFile(aktivitasFilePath);
                 Files.createFile(faseFilePath);
-                Files.writeString(autoIncFilePath, "0,0,0");
+                Files.writeString(autoIncFilePath, "0,0,0,0");
             } catch (IOException e) {
                 throw e;
             }
